+++ zetaclient/evm/inbounds.go
@@ -13,6 +13,8 @@ import (
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/zeta-chain/protocol-contracts/pkg/contracts/evm/erc20custody.sol"
 	"github.com/zeta-chain/protocol-contracts/pkg/contracts/evm/zetaconnector.non-eth.sol"
+	clientcommon "github.com/zeta-chain/zetacore/zetaclient/common"
+	"github.com/zeta-chain/zetacore/zetaclient/config"
 	clienttypes "github.com/zeta-chain/zetacore/zetaclient/types"
 
 	ethcommon "github.com/ethereum/go-ethereum/common"
@@ -296,6 +298,19 @@ func (ob *ChainClient) GetTransactionSender(tx *ethtypes.Transaction, blockHash 
 }
 
 func (ob *ChainClient) GetInboundVoteMsgForDepositedEvent(event *erc20custody.ERC20CustodyDeposited, sender ethcommon.Address) *types.MsgVoteOnObservedInboundTx {
+	// compliance check
+	maybeReceiver := ""
+	parsedAddress, _, err := common.ParseAddressAndData(hex.EncodeToString(event.Message))
+	if err == nil && parsedAddress != (ethcommon.Address{}) {
+		maybeReceiver = parsedAddress.Hex()
+	}
+	if config.ContainRestrictedAddress(sender.Hex(), clienttypes.BytesToEthHex(event.Recipient), maybeReceiver) {
+		clientcommon.PrintComplianceLog(ob.logger.ExternalChainWatcher, ob.logger.Compliance,
+			false, ob.chain.ChainId, event.Raw.TxHash.Hex(), sender.Hex(), clienttypes.BytesToEthHex(event.Recipient), "ERC20")
+		return nil
+	}
+
+	// donation check
 	if bytes.Equal(event.Message, []byte(DonationMessage)) {
 		ob.logger.ExternalChainWatcher.Info().Msgf("thank you rich folk for your donation! tx %s chain %d", event.Raw.TxHash.Hex(), ob.chain.ChainId)
 		return nil
@@ -329,6 +344,15 @@ func (ob *ChainClient) GetInboundVoteMsgForZetaSentEvent(event *zetaconnector.Ze
 		return nil
 	}
 	destAddr := clienttypes.BytesToEthHex(event.DestinationAddress)
+
+	// compliance check
+	sender := event.ZetaTxSenderAddress.Hex()
+	if config.ContainRestrictedAddress(sender, destAddr, event.SourceTxOriginAddress.Hex()) {
+		clientcommon.PrintComplianceLog(ob.logger.ExternalChainWatcher, ob.logger.Compliance,
+			false, ob.chain.ChainId, event.Raw.TxHash.Hex(), sender, destAddr, "Zeta")
+		return nil
+	}
+
 	if !destChain.IsZetaChain() {
 		paramsDest, found := ob.coreContext.GetEVMChainParams(destChain.ChainId)
 		if !found {
@@ -343,13 +367,13 @@ func (ob *ChainClient) GetInboundVoteMsgForZetaSentEvent(event *zetaconnector.Ze
 	}
 	message := base64.StdEncoding.EncodeToString(event.Message)
 	ob.logger.ExternalChainWatcher.Info().Msgf("ZetaSent inTx detected on chain %d tx %s block %d from %s value %s message %s",
-		ob.chain.ChainId, event.Raw.TxHash.Hex(), event.Raw.BlockNumber, event.ZetaTxSenderAddress.Hex(), event.ZetaValueAndGas.String(), message)
+		ob.chain.ChainId, event.Raw.TxHash.Hex(), event.Raw.BlockNumber, sender, event.ZetaValueAndGas.String(), message)
 
 	return zetabridge.GetInBoundVoteMessage(
-		event.ZetaTxSenderAddress.Hex(),
+		sender,
 		ob.chain.ChainId,
 		event.SourceTxOriginAddress.Hex(),
-		clienttypes.BytesToEthHex(event.DestinationAddress),
+		destAddr,
 		destChain.ChainId,
 		sdkmath.NewUintFromBigInt(event.ZetaValueAndGas),
 		message,
@@ -364,16 +388,27 @@ func (ob *ChainClient) GetInboundVoteMsgForZetaSentEvent(event *zetaconnector.Ze
 }
 
 func (ob *ChainClient) GetInboundVoteMsgForTokenSentToTSS(tx *ethtypes.Transaction, sender ethcommon.Address, blockNumber uint64) *types.MsgVoteOnObservedInboundTx {
+	message := hex.EncodeToString(tx.Data())
+
+	// compliance check
+	maybeReceiver := ""
+	parsedAddress, _, err := common.ParseAddressAndData(message)
+	if err == nil && parsedAddress != (ethcommon.Address{}) {
+		maybeReceiver = parsedAddress.Hex()
+	}
+	if config.ContainRestrictedAddress(sender.Hex(), maybeReceiver) {
+		clientcommon.PrintComplianceLog(ob.logger.ExternalChainWatcher, ob.logger.Compliance,
+			false, ob.chain.ChainId, tx.Hash().Hex(), sender.Hex(), sender.Hex(), "Gas")
+		return nil
+	}
+
+	// donation check
 	if bytes.Equal(tx.Data(), []byte(DonationMessage)) {
 		ob.logger.ExternalChainWatcher.Info().Msgf("thank you rich folk for your donation! tx %s chain %d", tx.Hash().Hex(), ob.chain.ChainId)
 		return nil
 	}
-	message := ""
-	if len(tx.Data()) != 0 {
-		message = hex.EncodeToString(tx.Data())
-	}
 	ob.logger.ExternalChainWatcher.Info().Msgf("TSS inTx detected on chain %d tx %s block %d from %s value %s message %s",
-		ob.chain.ChainId, tx.Hash().Hex(), blockNumber, sender.Hex(), tx.Value().String(), hex.EncodeToString(tx.Data()))
+		ob.chain.ChainId, tx.Hash().Hex(), blockNumber, sender.Hex(), tx.Value().String(), message)
 
 	return zetabridge.GetInBoundVoteMessage(
 		sender.Hex(),